Split tests file into logical pieces
